Move SignalBars to UI directory
